requirements.txt
fastapi
uvicorn
python-multipart
pinecone-client
openai
redis
pymongo
pdfplumber
python-dotenv
